Improve performance of `iterdirs` by pre-filling set
`unique_everseen` places values in a set by
calling  `set.add` on every new value that
it encounters.  This is palpably slower than
passing known values to the constructor,
especially when working with zips which contain
many files or sub-directories.

Taking CPython as an example - prior to this change:

    >>> zip = zipp.Path('cpython.zip')
    >>> len(zip.root.namelist())
    4575
    >>> timeit.timeit(lambda: list(zip.iterdir()), number=10)
    3.090643104999799

After:

    >>> timeit.timeit(lambda: list(zip.iterdir()), number=10)
    0.44296542000000727