Memoize the zipfile namelist when it's safe to do so.
If the zipfile is in 'r' mode, we know it can't change under
us, so it's safe to lazily memoize the namelist (augmented with
the implied directories).

We memoize it as a set, to make lookups into it efficient.
This emphasizes that iterdir's iteration order is arbitrary.

This was actually true before - the order we were getting in
practice was "all the explicit names, in zipfile insertion order,
followed by all the implicit dirs, in zipfile insertion order of
the descendants from which we inferred them). This isn't a
particularly intuitive or useful order for end users. However
pathlib's iterdir, which this one emulates, also does not guarantee
anything about order (it appears to in practice on MacOS, but not
on Linux). So it seems right to double-down on this lack of order.

This change adds ordering semantics for Path objects, to make it
easy to sort them if you need to. The tests now use this sorting.

As a result of this change, in 'r' mode, exists() and joinpath()
are constant time instead of linear time.  They are still linear
in other modes.