Add dependency on `pathlib-abc`
Make `zipp.Path` subclass `pathlib_abc.ReadablePath`. This allows us to
remove implementations of `name`, `suffix`, `suffixes`, `stem`,
`read_text()`, `read_bytes()`, `glob()` and `joinpath()`, and to simplify
implementations of a couple of other methods.

Maintain a tree of `PathInfo` objects representing the hierarchy of zip
file members. We traverse the tree whenever we need to resolve a path to a
`ZipInfo` object. This effectively hides the `.zip`-specific quirk that
directories are recorded with a trailing slash in their filenames.