Fix `TypeError` from `str(Path(...))`
Make `Path.__str__()` return `self.at` rather than attempting to join the
member path onto the zip file path.

When the `ZipFile` is named, the new behaviour produces a more
straightforward string representation that is less likely to be mistaken
for a real filesystem path.

When the `ZipFile` is unnamed, the new behaviour ensures that a `TypeError`
isn't raised.

Resolves #134