Add dependency on `pathlib-abc`
Make `zipp.Path` subclass `pathlib_abc.ReadablePath`. This allows us to
remove implementations of `read_text()`, `read_bytes()`, `glob()`,
`joinpath()` and `__truediv__()`, and to simplify implementations of a
couple of few more methods.

Maintain a tree of `PathInfo` objects representing the hierarchy of zip
file members. We traverse the tree whenever we need to resolve a path to
a `ZipInfo` object. This effectively hides the `.zip`-specific quirk that
directories are recorded with a trailing slash in their filenames.

Adjust `__str__()` so that it doesn't raise for unnamed zip files. Instead
it returns a string beginning with `:fileobj:`. The prefix is made
available as a new `anchor` attribute.

Add the following methods/attributes, which are required by `ReadablePath`:

- `info`
- `parser`
- `__open_rb__()`
- `readlink()`
- `with_segments()`

Disable the following `ReadablePath` methods/attributes that we don't (yet) test in zipp:

- `anchor`
- `parts`
- `parents`
- `__rtruediv__()`
- `with_name()`
- `with_stem()`
- `with_suffix()`
- `full_match()`
- `walk()`
- `copy()`
- `copy_into()`